refactor(components): update import paths for type definitions
Replaced relative imports of type definitions with more explicit paths to improve code clarity and maintainability.
